refactor: extract addCommentViaUI into shared helpers
Deduplicate the addCommentViaUI helper that was copied across 6 test
files. The shared version supports an opts object with lineIndex and
waitText parameters for flexibility.
